redhat/Makefile: honor BREW_FLAGS like the kernel
When doing make rh-brew on a RHEL kernel one can set BREW_FLAGS
to anything that 'brew build' accepts (see 'brew build -h'). This
is useful to avoid wasting resources when only one arch needs to
be compiled for a test, as the arch-override parameter can be
set. For example,

  BREW_FLAGS='--arch-override=aarch64' make -C redhat rh-brew

only builds for aarch64.

Signed-off-by: Andrew Jones <drjones@redhat.com>
Signed-off-by: Miroslav Rezanina <mrezanin@redhat.com>