Avoid assertion in check_interrupts_property()
If running on a tree with an 'interrupt-parent' property which contains
an invalid phandle (0 or -1, not merely for a node which doesn't exist),
then check_interrupts_property() will trip the assertion in
get_node_by_phandle().

There's logic that almost detects this, but it only handles the overlay
case, where we can't fully check because the links will be fixed up later.

For the non-overlay case, this is definitely a bad property, but we
shouldn't crash.  Fix it by failing the check early.

Fixes: c1e55a5513e9 ("checks: fix handling of unresolved phandles for dts plugins")
Fixes: ee3d26f6960b ("checks: add interrupts property check")
Signed-off-by: David Gibson <david@gibson.dropbear.id.au>